mpi: refactor to support stacking with MPI.
Introduces MpiMultiplexTaskRunner, which provides for simple,
embarassingly parallel processing over MPI.  The user shouldn't have to
change anything in the Task to take advantage of the functionality, except
ensuring that any expensive argument parsing (e.g., reading Wcs for all
input images) is minimised.  To this end, the MpiArgumentParser now allows
specification as to whether a DataIdContainer is defined only on the root
(for expensive parsing) or not.