pool: add namespaces ("context") for cache/store
Different layers of a program may want to use the pool cache/store,
but this would result in corruption.  Adding a namespace (called
"context") fixes this problem.
The Pool class now takes just a context (there's no need to allow
a different comm or root, because that's already set in stone by
the time the user creates a Pool), and is a wrapper around
PoolMaster.