BatchCmdLineTask: apply overrides from Task when parsing arguments
Because we're now saving the config before submitting to the batch
system, the config we construct must be identical to what is constructed
when the Task runs (and it could lead to other trouble if it was not),
or the Task will fail to validate the config when it starts up on the
batch system.  We were missing an important element of the config
overrides --- those applied within the Task's applyOverrides method.