batch: add --batch-stats argument to print process statistics on completion
This adds a Linux-specific feature for gathering and printing the process
statistics atexit.  On Linux, the /proc/self/status file contains key-value
pairs with statistics about the process.  By printing this file at the end
of a run, we can check that the peak memory usage (VmPeak) and other
quantities are reasonable.