parallel: use --cores to specify number of cores
For some slurm systems and for SMP, the only thing that matters is
the total number of cores requested, rather than a combination
of nodes and number of processors per node.  Added support for
specifying the number of cores (via --cores).  Made this optional
for SMP (supporting the old --nodes=1 --procs=NN as a synonym for
--cores=NN), optional for slurm, and disallowed for PBS.

Cherry-picked from hscPipeBase commit fa131a6 (HSC-1369).