Add script that demonstrates and tests BatchCmdLineTask
The Slurm, PBS, etc. functionality can't be tested with the
usual unit tests, but we can make it easy for the developer
to exercise things. Here we add a very simple demonstration
script that is intended to show how BatchCmdLineTask can
be used, and exercise it to ensure it's working.