Comm: add alternate version of scatter
The default version apparently pickles the entire 'dataList', which
can cause errors if the pickle size grows over 2^31 bytes due to
fundamental problems with pickle in python 2 [1][2] (causing, e.g.,
"OverflowError: integer 2155421250 does not fit in 'int'"). Instead,
we send the data to each slave node in turn; this reduces the pickle
size.

[1] http://bugs.python.org/issue11564
[2] https://www.python.org/dev/peps/pep-3154/