Pool.reduce: collect results individually when not reducing
DM-5989 changed the behaviour for Pool.map (implemented by
Pool.reduce with reducer=None) to collect the results of its operations
and then send them all at the end.  However, this makes it easier to
exceed the 32-bit size limit for pickles in python 2, which brings the
entire process down. This patch implements the behaviour prior to
DM-5989, which is to send each result as it becomes available. We hope
that by dividing the N results into N separate reports, we can keep
under the pickle size limit.

Note that reduceNoBalance does not have this feature, but it has
never had it and we haven't had problems; we'll fix that if it
becomes necessary.